refactor: fix variable name bug, remove unused collection
