--- src/models/schemas/task.js
@@ -1,34 +1,0 @@
-const { Schema, model } = require("mongoose");
-
-const TaskSchema = new Schema({
-    name: {
-        type: String,
-        required: true,
-    },
-    points: {
-        type: Number,
-        required: true,
-    },
-    description: {
-        type: String,
-    },
-    createdAt: {
-        type: Date,
-        required: true,
-    },
-    // 最多可執行次數 / 0 代表無限
-    maxRunCount: {
-        type: Number,
-        required: true,
-    },
-});
-
-// create index here
-// TaskSchema.index({});
-
-const Task = model("Task", TaskSchema, "tasks");
-
-module.exports = {
-    TaskSchema,
-    Task,
-};
