--- src/models/schemas/userModel.js
@@ -1,6 +1,0 @@
-const mongoose = require("mongoose");
-const userSchema = require("./userSchema");
-
-const User = mongoose.model("User", userSchema);
-
-module.exports = User;
